+++ tests/TestCase.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace ArtisanToolbox\Core\Tests;
 
 use ArtisanToolbox\Core\CoreServiceProvider;
+use Inertia\ServiceProvider as InertiaServiceProvider;
 use Orchestra\Testbench\TestCase as Orchestra;
 
 abstract class TestCase extends Orchestra
@@ -12,6 +13,7 @@ abstract class TestCase extends Orchestra
     protected function getPackageProviders($app): array
     {
         return [
+            InertiaServiceProvider::class,
             CoreServiceProvider::class,
         ];
     }